@@ -1,15 +1,17 @@
-import React, { KeyboardEvent, useCallback, useState } from 'react';
-import { GetServerSideProps } from 'next';
-import { useRouter } from 'next/router';
-import styles from 'styles/music.module.scss';
-import Dropdown from 'components/dropdown';
-import KEY_EVENTS from 'utils/events';
-import classNames from 'utils/classNames';
-import { getAccessToken, getUniqueAlbums, getTopTracks, Album } from 'utils/spotifyHttpClient';
+'use client';
 
-type MusicProps = {
-  topAlbums: Album[];
+import { KeyboardEvent, useCallback, useState } from 'react';
+import { usePathname, useRouter } from 'next/navigation';
+import Image from 'next/image';
+import styles from '@/styles/music.module.scss';
+import Dropdown from '@/components/Dropdown';
+import KEY_EVENTS from '@/utils/events';
+import classNames from '@/utils/classNames';
+import { Album } from '@/types/music';
+
+type MusicGridProps = {
   duration: Duration;
+  topAlbums: Album[];
 };
 
 type Duration = 'short' | 'medium' | 'long';
@@ -20,8 +22,9 @@ const durationLabels = {
   long: 'years'
 };
 
-const Music = ({ topAlbums, duration = 'medium' }: MusicProps) => {
+const MusicGrid = ({ duration, topAlbums }: MusicGridProps) => {
   const router = useRouter();
+  const pathname = usePathname();
   const [openIndex, setOpenIndex] = useState<number>();
 
   const onAlbumClick = useCallback(
@@ -44,7 +47,7 @@ const Music = ({ topAlbums, duration = 'medium' }: MusicProps) => {
   const onDurationSelect = (option: string) => {
     setOpenIndex(-1);
     const duration = Object.keys(durationLabels).find(key => durationLabels[key as Duration] === option);
-    router.push(`${router.pathname}?duration=${duration}`);
+    router.push(`${pathname}?duration=${duration}`);
   };
 
   return (
@@ -67,13 +70,14 @@ const Music = ({ topAlbums, duration = 'medium' }: MusicProps) => {
                 <span>{artist}</span>
                 <span>{name}</span>
               </div>
-              <img
+              <Image
                 className={classNames([styles.albumArt, openIndex === index && styles.open])}
                 onKeyDown={event => onAlbumKeyDown(event, index)}
                 tabIndex={0}
                 alt={`${name} by ${artist}`}
                 loading="lazy"
                 src={images[0].url}
+                fill
               />
             </div>
           );
@@ -83,20 +87,4 @@ const Music = ({ topAlbums, duration = 'medium' }: MusicProps) => {
   );
 };
 
-export const getServerSideProps: GetServerSideProps = async ({ query }) => {
-  const { SPOTIFY_REFRESH_TOKEN, SPOTIFY_CLIENT_ID, SPOTIFY_CLIENT_SECRET } = process.env;
-
-  const durationQuery = Array.isArray(query.duration) ? query.duration[0] : query.duration;
-  const durationValue = durationQuery || 'medium';
-  const duration = ['short', 'medium', 'long'].includes(durationValue) ? durationValue : 'medium';
-
-  if (SPOTIFY_REFRESH_TOKEN && SPOTIFY_CLIENT_ID && SPOTIFY_CLIENT_SECRET) {
-    const accessToken = await getAccessToken(SPOTIFY_REFRESH_TOKEN, SPOTIFY_CLIENT_ID, SPOTIFY_CLIENT_SECRET);
-    const topTracks = await getTopTracks(accessToken, duration);
-    const topAlbums = getUniqueAlbums(topTracks);
-    return { props: { topAlbums, duration } };
-  }
-  return { props: { topAlbums: [], duration } };
-};
-
-export default Music;
+export default MusicGrid;
